c2s {remoter}	R Documentation

Client-to-Server Object Transfer

Description

This function allows you to pass an object from the local R session (the client) to server.

Usage

c2s(object, newname, env = .GlobalEnv)

Arguments

object
A local R object.

newname
The name the object should take when it is stored on the remote server. If left blank, the remote name will be the same as the original (local) object's name.

env
The environment into which the assignment will take place. The default is the remoter "working environment".

Details

Localize R objects.

Value

Returns TRUE invisibly on successful exit.

Examples

## Not run: 
### Prompts are listed to clarify when something is eval'd locally vs remotely
> library(remoter)
> x <- "some data"
> remoter::connect("my.remote.server")
remoteR> x
### Error: object 'x' not found
remoteR> c2s(x)
remoteR> x
###  [1] "some data" 

## End(Not run)

[Package remoter version 0.3-1 Index]