evalc {remoter}	R Documentation

evalc

Description

A function to evaluate expressions on the client's R session. To eval expressions on the server, just use eval(). Instead of using this function, you could also just kill the client, do your local operations, then re-run your client() command.

Usage

evalc(expr)

Arguments

expr
Expression to be evaluated on the client.

Details

Evaluate expressions on the client.

Value

Returns TRUE invisibly on successful exit.

[Package remoter version 0.3-1 Index]