exit {remoter}	R Documentation

exit

Description

This function cleanly shuts down the remoter server the client is currently connected to, as well as shutting down the client. One can also use q() (while the client is running), and this will not close the active R session on the client.

Usage

exit(client.only = TRUE, q.server = TRUE)

Arguments

client.only
Logical; if TRUE, then the client disconnects from the server. Otherwise, the server is shut down together with the client.

q.server
Logical; if TRUE, then the server calls q("no") after shuting down with the client. This is useful for cases where the server is running in an interactive R session, and you wish to shut the entire thing down.

Details

Exit the remoter client/server.

Value

Returns TRUE invisibly on successful exit.

[Package remoter version 0.3-1 Index]