has.sodium {remoter}	R Documentation

has.sodium

Description

Report if the sodium package is availabe for use.

Usage

has.sodium()

Value

Returns TRUE if the sodium package is available, and FALSE otherwise.

[Package remoter version 0.3-0 Index]