is.secure {remoter}	R Documentation

is.secure

Description

Report if communications with the connected server are encrypted.

Usage

is.secure()

Value

Returns TRUE if messages between client and server are currently encrypted, and FALSE if not. If the client is not currently running (i.e., if executed from just a regular R prompt), then NA is returned.

[Package remoter version 0.3-1 Index]