lsc {remoter}	R Documentation

ls on Client

Description

A function to view environments on the client's R session. To view objects on the server, just use ls(). Instead of using this function, you could also just kill the client, do your local operations, then re-run your client() command.

Usage

lsc(envir, all.names = FALSE, pattern)

Arguments

envir
Environment (as in ls()).

all.names
Logical that determines if all names are returned or those beginning with a '.' are omitted (as in ls()).

pattern
Optional regular expression (as in ls()).

Details

View objects on the client.

Value

Returns TRUE invisibly on successful exit.

[Package remoter version 0.3-0 Index]