remoter-package {remoter}	R Documentation

remoter

Description

A package for running R commands on a remote R session. With remoter, you can use RStudio running on your laptop to execute commands on an R session running on, for example, Amazon's EC2.

Author(s)

Drew Schmidt and Wei-Chen Chen

References

Project URL: https://github.com/wrathematics/remoter

[Package remoter version 0.3-0 Index]