rmc {remoter}	R Documentation

rmc

Description

A function to remove objects from the client's R session. To remove objects on the server, just use rm(). Instead of using this function, you could also just kill the client, do your local operations, then re-run your client() command.

Usage

rmc(..., list = character(), envir)

Arguments

...
Objects to be removed from the client's R session.

list
Character vector naming objects to be removed (as in rm()).

envir
Environment (as in rm()).

Details

Remove objects on the client.

Value

Returns TRUE invisibly on successful exit.

[Package remoter version 0.3-1 Index]