s2c {remoter}	R Documentation

Server-to-Client Object Transfer

Description

This function allows you to pass an object from the server to the local R session behind the client.

Usage

s2c(object, newname, env = .GlobalEnv)

Arguments

object
A remote R object.

newname
The name the object should take when it is stored on the local client's R session. If left blank, the local name will be the same as the original (remote) object's name.

env
The environment into which the assignment will take place. The default is the global environment.

Details

Localize R objects.

Value

Returns TRUE invisibly on successful exit.

Examples

## Not run: 
### Prompts are listed to clarify when something is eval'd locally vs remotely
> library(remoter)
> y
###  Error: object 'y' not found
> remoter::connect("my.remote.server")
remoteR> x
### Error: object 'x' not found
remoteR> x <- "some data"
remoteR> x
###  [1] "some data" 
remoteR> s2c(x, "y")
remoteR> q()
> y
###  [1] "some data"

## End(Not run)

[Package remoter version 0.3-0 Index]