server {remoter}	R Documentation

Server Launcher

Description

Launcher for the remoter server.

Usage

server(port = 55555, password = NULL, maxretry = 5,
  secure = has.sodium(), log = TRUE, verbose = FALSE, showmsg = FALSE)

Arguments

port
The port (number) that will be used for communication between the client and server. The port value for the client and server must agree.

password
A password the client must enter before the user can process commands on the server. If the value is NULL, then no password checking takes place.

maxretry
The maximum number of retries for passwords before shutting everything down.

secure
Logical; enables encryption via public key cryptography of the 'sodium' package is available.

log
Logical; enables some basic logging in the server.

verbose
Logical; enables the verbose logger.

showmsg
Logical; if TRUE, messages from the client are logged

Details

By a 'secure' server, we mean one that encrypts messages it sends and only accepts encrypted messages. Encryption uses public key cryptography, using the 'sodium' package.

If the 'sodium' package is available to the server, then by default the server will be secure. If the package is not available, then you will not be able to start a secure server. If the server is secure, then a client can only connect if the client has the 'sodium' package available.

Value

Returns TRUE invisibly on successful exit.

[Package remoter version 0.3-1 Index]